% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scatterMatrix.R
\name{renderScatterMatrix}
\alias{renderScatterMatrix}
\title{Widget render function for use in Shiny}
\usage{
renderScatterMatrix(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

