% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scatterMatrix.R
\name{scatterMatrix}
\alias{scatterMatrix}
\title{D3 scatterplot matrix}
\usage{
scatterMatrix(data = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{data.frame containing the data}

\item{width}{}

\item{height}{}
}
\description{
This is a wrapper around the D3 scatterplot matrix, originally developed by
Mike Bostock and extended by Benjie Chen.
}

